Revert "orginize project structure like a set of modules"
This reverts commit af3b4a40be600ef91ee223ac7a4cfffc69841810.

+++ src/org/mql/platform/models/Internship.java
@@ -4,7 +4,6 @@ import java.time.LocalDate;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
-
 import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
@@ -12,8 +11,6 @@ import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
-
-import org.mql.platform.commons.models.Document;
 
 /**
  * @author chermehdi
@@ -77,11 +74,11 @@ public class Internship extends Experiment {
     this.endDate = endDate;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
